python-animation-app/requirements.txt
Pillow
imageio
numpy
matplotlib